Add process interruption and stop button to editor
Introduced a mechanism to interrupt running processes in the flow editor by adding an InterruptProcess method to Network and a corresponding static flag in Node. Added a stop button with an associated image and event handler in the UI to allow users to halt execution. Also included a sample endless loop flowchart for testing.